+++ ts/src/common.ts
@@ -4,8 +4,21 @@ import {
   InstalledAppInfo,
   InstalledCell,
 } from "@holochain/client";
-import { AgentHapp, CellZomeCallRequest, IConductor } from "./types.js";
+import {
+  AgentHapp,
+  CallableCell,
+  CellZomeCallRequest,
+  IConductor,
+} from "./types.js";
 
+/**
+ * Add all agents of all conductors to each other. Shortcuts peer discovery
+ * through a bootstrap server or gossiping.
+ *
+ * @param conductors - Conductors to mutually exchange all agents with.
+ *
+ * @public
+ */
 export const addAllAgentsToAllConductors = async (conductors: IConductor[]) => {
   await Promise.all(
     conductors.map(async (playerToShareAbout, playerToShareAboutIdx) => {
@@ -80,3 +93,21 @@ const getCallableCell = (
     return callZomeResponse;
   },
 });
+
+/**
+ * Get a shorthand function to call a Cell's Zome.
+ *
+ * @param cell - The Cell to call the Zome on.
+ * @param zomeName - The name of the Zome to call.
+ * @returns A function to call the specified Zome.
+ *
+ * @public
+ */
+export const getZomeCaller =
+  (cell: CallableCell, zomeName: string) =>
+  <T>(fnName: string, payload: unknown): Promise<T> =>
+    cell.callZome<T>({
+      zome_name: zomeName,
+      fn_name: fnName,
+      payload,
+    });